Changing data item random property id to smallint

+++ Backend/src/modules/data/domain_value/item_random_property.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone, Serialize, JsonSchema)]
 pub struct ItemRandomProperty {
   pub expansion_id: u8,
-  pub id: u32,
+  pub id: u16,
   pub localization_id: u32,
   pub enchant_ids: Vec<u32>
 }
